# nodejs-server-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is Swagger UI environment generated for the TMF Service Catalog\
    \ Management specification"
  version: "2.0.0"
  title: "Service Catalog Management"
host: "localhost:8080"
basePath: "/tmf-api/serviceCatalogManagement/v2/"
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /serviceCandidate:
    get:
      tags:
      - "serviceCandidate"
      summary: "List or find 'ServiceCandidate' objects"
      operationId: "listServiceCandidate"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma separated properties to display in response"
        required: false
        type: "string"
      - name: "name"
        in: "query"
        description: "For filtering: Name given to this REST resource"
        required: false
        type: "string"
      - name: "type"
        in: "query"
        description: "For filtering: Class type of this REST resource"
        required: false
        type: "string"
      - name: "@schemaLocation"
        in: "query"
        description: "For filtering: This field provides a link to the schema describing\
          \ this REST resource"
        required: false
        type: "string"
      - name: "@baseType"
        in: "query"
        description: "For filtering: The (immediate) base class type of this REST\
          \ resource"
        required: false
        type: "string"
      - name: "version"
        in: "query"
        description: "For filtering: the version of service candidate"
        required: false
        type: "string"
      - name: "validForStartDateTime"
        in: "query"
        description: "For filtering: An instant of time, starting at the TimePeriod"
        required: false
        type: "string"
        format: "date-time"
      - name: "validForEndDateTime"
        in: "query"
        description: "For filtering: An instant of time, ending at the TimePeriod."
        required: false
        type: "string"
        format: "date-time"
      - name: "lastUpdate"
        in: "query"
        description: "For filtering: Date and time of the last update of this REST\
          \ resource"
        required: false
        type: "string"
        format: "date-time"
      - name: "lifecycleStatus"
        in: "query"
        description: "For filtering: Used to indicate the current lifecycle status\
          \ of the service candidate."
        required: false
        type: "string"
      - name: "categoryVersion"
        in: "query"
        description: "For filtering: Category version"
        required: false
        type: "string"
      - name: "categoryName"
        in: "query"
        description: "For filtering: Name of the category"
        required: false
        type: "string"
      - name: "categoryType"
        in: "query"
        description: "For filtering: the class type of the category"
        required: false
        type: "string"
      - name: "serviceSpecificationVersion"
        in: "query"
        description: "For filtering: ServiceSpecification version"
        required: false
        type: "string"
      - name: "serviceSpecificationName"
        in: "query"
        description: "For filtering: Name given to the ServiceSpecification"
        required: false
        type: "string"
      - name: "serviceSpecificationType"
        in: "query"
        description: "For filtering: The (class) type of the ServiceSpecification"
        required: false
        type: "string"
      responses:
        200:
          description: "Ok"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceCandidate"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCandidate"
    post:
      tags:
      - "serviceCandidate"
      summary: "Creates a 'ServiceCandidate'"
      operationId: "createServiceCandidate"
      parameters:
      - in: "body"
        name: "serviceCandidate"
        description: "The Service Candidate to be created"
        required: true
        schema:
          $ref: "#/definitions/ServiceCandidate_Create"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/ServiceCandidate"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCandidate"
  /serviceCandidate/{id}:
    get:
      tags:
      - "serviceCandidate"
      summary: "Retrieves a 'ServiceCandidate' by Id"
      operationId: "retrieveServiceCandidate"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Service Candidate"
        required: true
        type: "string"
      responses:
        200:
          description: "Ok"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ServiceCandidate"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCandidate"
    delete:
      tags:
      - "serviceCandidate"
      summary: "Deletes a 'ServiceCandidate' by Id"
      operationId: "deleteServiceCandidate"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Service Candidate"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCandidate"
    patch:
      tags:
      - "serviceCandidate"
      summary: "Updates partially a 'ServiceCandidate' by Id"
      operationId: "patchServiceCandidate"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Service Candidate"
        required: true
        type: "string"
      - in: "body"
        name: "serviceCandidate"
        description: "The Service Candidate to be updated"
        required: true
        schema:
          $ref: "#/definitions/ServiceCandidate_Update"
      responses:
        201:
          description: "Updated"
          schema:
            $ref: "#/definitions/ServiceCandidate"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ServiceCandidate"
...
  /importJob:
    get:
      tags:
      - "importJob"
      summary: "List or find 'ImportJob' objects"
      operationId: "listImportJob"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma separated properties to display in response"
        required: false
        type: "string"
      - name: "contentType"
        in: "query"
        description: "For filtering: Indicates the format of the imported data"
        required: false
        type: "string"
      - name: "path"
        in: "query"
        description: "For filtering: URL of the root resource where the content of\
          \ the file specified by the import job must be applied"
        required: false
        type: "string"
      - name: "status"
        in: "query"
        description: "For filtering: Status of the import job (not started, running,\
          \ succeeded, failed)"
        required: false
        type: "string"
      - name: "url"
        in: "query"
        description: "For filtering: URL of the file containing the data to be imported"
        required: false
        type: "string"
      - name: "completionDate"
        in: "query"
        description: "For filtering: Date at which the job was completed"
        required: false
        type: "string"
        format: "date-time"
      - name: "creationDate"
        in: "query"
        description: "For filtering: Date at which the job was created"
        required: false
        type: "string"
        format: "date-time"
      - name: "errorLog"
        in: "query"
        description: "For filtering: Reason for failure if status is failed"
        required: false
        type: "string"
      responses:
        200:
          description: "Ok"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ImportJob"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ImportJob"
    post:
      tags:
      - "importJob"
      summary: "Creates a 'ImportJob'"
      operationId: "createImportJob"
      parameters:
      - in: "body"
        name: "importJob"
        description: "The Import Job to be created"
        required: true
        schema:
          $ref: "#/definitions/ImportJob_Create"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/ImportJob"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ImportJob"
  /importJob/{id}:
    get:
      tags:
      - "importJob"
      summary: "Retrieves a 'ImportJob' by Id"
      operationId: "retrieveImportJob"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Import Job"
        required: true
        type: "string"
      responses:
        200:
          description: "Ok"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ImportJob"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ImportJob"
    delete:
      tags:
      - "importJob"
      summary: "Deletes a 'ImportJob' by Id"
      operationId: "deleteImportJob"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Import Job"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ImportJob"
  /exportJob:
    get:
      tags:
      - "exportJob"
      summary: "List or find 'ExportJob' objects"
      operationId: "listExportJob"
      parameters:
      - name: "fields"
        in: "query"
        description: "Comma separated properties to display in response"
        required: false
        type: "string"
      - name: "query"
        in: "query"
        description: "For filtering: Used to scope the exported data"
        required: false
        type: "string"
      - name: "path"
        in: "query"
        description: "For filtering: URL of the root resource acting as the source\
          \ for streaming content to the file specified by the export job"
        required: false
        type: "string"
      - name: "contentType"
        in: "query"
        description: "For filtering: The format of the exported data"
        required: false
        type: "string"
      - name: "status"
        in: "query"
        description: "For filtering: Status of the export job (not started, running,\
          \ succeeded, failed)"
        required: false
        type: "string"
      - name: "url"
        in: "query"
        description: "For filtering: URL of the file containing the data to be exported"
        required: false
        type: "string"
      - name: "completionDate"
        in: "query"
        description: "For filtering: Data at which the job was completed"
        required: false
        type: "string"
        format: "date-time"
      - name: "creationDate"
        in: "query"
        description: "For filtering: Date at which the job was created"
        required: false
        type: "string"
        format: "date-time"
      - name: "errorLog"
        in: "query"
        description: "For filtering: Reason for failure"
        required: false
        type: "string"
      responses:
        200:
          description: "Ok"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ExportJob"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ExportJob"
    post:
      tags:
      - "exportJob"
      summary: "Creates a 'ExportJob'"
      operationId: "createExportJob"
      parameters:
      - in: "body"
        name: "exportJob"
        description: "The Export Job to be created"
        required: true
        schema:
          $ref: "#/definitions/ExportJob_Create"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/ExportJob"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ExportJob"
  /exportJob/{id}:
    get:
      tags:
      - "exportJob"
      summary: "Retrieves a 'ExportJob' by Id"
      operationId: "retrieveExportJob"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Export Job"
        required: true
        type: "string"
      responses:
        200:
          description: "Ok"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/ExportJob"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ExportJob"
    delete:
      tags:
      - "exportJob"
      summary: "Deletes a 'ExportJob' by Id"
      operationId: "deleteExportJob"
      parameters:
      - name: "id"
        in: "path"
        description: "Identifier of the Export Job"
        required: true
        type: "string"
      responses:
        204:
          description: "Deleted"
        400:
          description: "Bad Request"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "ExportJob"
...
definitions:
...
  ServiceCandidate:
    type: "object"
    properties:
      id:
        type: "string"
        description: "Unique identifier of this REST resource"
      href:
        type: "string"
        description: "Hyperlink reference to this REST resource"
      name:
        type: "string"
        description: "Name given to this REST resource"
      description:
        type: "string"
        description: "Description of this REST resource"
      '@type':
        type: "string"
        description: "Class type of this REST resource"
      '@schemaLocation':
        type: "string"
        description: "This field provides a link to the schema describing this REST\
          \ resource"
      '@baseType':
        type: "string"
        description: "The (immediate) base class type of this REST resource"
      version:
        type: "string"
        description: "the version of service candidate"
      validFor:
        description: "The period for which this REST resource is valid"
        $ref: "#/definitions/TimePeriod"
      lastUpdate:
        type: "string"
        format: "date-time"
        description: "Date and time of the last update of this REST resource"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status of the service\
          \ candidate."
      category:
        type: "array"
        items:
          $ref: "#/definitions/CategoryRef"
      serviceSpecification:
        $ref: "#/definitions/ServiceSpecificationRef"
    description: "ServiceCandidate is an entity that makes a service specification\
      \ available to a catalog. A\nServiceCandidate and its associated service specification\
      \ may be published - made visible - in any number of service catalogs, or in\
      \ none. One service specification can be composed of other service specifications."
    example:
      lifecycleStatus: "lifecycleStatus"
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      '@type': "@type"
      serviceSpecification:
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        version: "version"
      description: "description"
      version: "version"
      '@baseType': "@baseType"
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      id: "id"
      href: "href"
      '@schemaLocation': "@schemaLocation"
      category:
      - name: "name"
        id: "id"
        href: "href"
        version: "version"
      - name: "name"
        id: "id"
        href: "href"
        version: "version"
  ServiceCandidate_Create:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
        description: "Name given to this REST resource"
      description:
        type: "string"
        description: "Description of this REST resource"
      '@type':
        type: "string"
        description: "Class type of this REST resource"
      '@schemaLocation':
        type: "string"
        description: "This field provides a link to the schema describing this REST\
          \ resource"
      '@baseType':
        type: "string"
        description: "The (immediate) base class type of this REST resource"
      version:
        type: "string"
        description: "the version of service candidate"
      validFor:
        description: "The period for which this REST resource is valid"
        $ref: "#/definitions/TimePeriod"
      lastUpdate:
        type: "string"
        format: "date-time"
        description: "Date and time of the last update of this REST resource"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status of the service\
          \ candidate."
      category:
        type: "array"
        items:
          $ref: "#/definitions/CategoryRef"
      serviceSpecification:
        $ref: "#/definitions/ServiceSpecificationRef"
    description: "ServiceCandidate is an entity that makes a service specification\
      \ available to a catalog. A\nServiceCandidate and its associated service specification\
      \ may be published - made visible - in any number of service catalogs, or in\
      \ none. One service specification can be composed of other service specifications.\n\
      Skipped properties: id,href"
    example:
      '@baseType': "@baseType"
      lifecycleStatus: "lifecycleStatus"
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      '@type': "@type"
      lastUpdate: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      serviceSpecification:
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        version: "version"
      description: "description"
      '@schemaLocation': "@schemaLocation"
      category:
      - name: "name"
        id: "id"
        href: "href"
        version: "version"
      - name: "name"
        id: "id"
        href: "href"
        version: "version"
      version: "version"
  ServiceCandidate_Update:
    type: "object"
    properties:
      name:
        type: "string"
        description: "Name given to this REST resource"
      description:
        type: "string"
        description: "Description of this REST resource"
      '@schemaLocation':
        type: "string"
        description: "This field provides a link to the schema describing this REST\
          \ resource"
      '@baseType':
        type: "string"
        description: "The (immediate) base class type of this REST resource"
      version:
        type: "string"
        description: "the version of service candidate"
      validFor:
        description: "The period for which this REST resource is valid"
        $ref: "#/definitions/TimePeriod"
      lifecycleStatus:
        type: "string"
        description: "Used to indicate the current lifecycle status of the service\
          \ candidate."
      category:
        type: "array"
        items:
          $ref: "#/definitions/CategoryRef"
      serviceSpecification:
        $ref: "#/definitions/ServiceSpecificationRef"
    description: "ServiceCandidate is an entity that makes a service specification\
      \ available to a catalog. A\nServiceCandidate and its associated service specification\
      \ may be published - made visible - in any number of service catalogs, or in\
      \ none. One service specification can be composed of other service specifications.\n\
      Skipped properties: id,href,id,href,@type,lastUpdate"
    example:
      '@baseType': "@baseType"
      lifecycleStatus: "lifecycleStatus"
      validFor:
        startDateTime: "2000-01-23T04:56:07.000+00:00"
        endDateTime: "2000-01-23T04:56:07.000+00:00"
      name: "name"
      serviceSpecification:
        '@type': "@type"
        name: "name"
        id: "id"
        href: "href"
        version: "version"
      description: "description"
      '@schemaLocation': "@schemaLocation"
      category:
      - name: "name"
        id: "id"
        href: "href"
        version: "version"
      - name: "name"
        id: "id"
        href: "href"
        version: "version"
      version: "version"
...
  ImportJob:
    type: "object"
    properties:
      id:
        type: "string"
        description: "Identifier of the import job"
      href:
        type: "string"
        description: "Reference of the import job"
      contentType:
        type: "string"
        description: "Indicates the format of the imported data"
      path:
        type: "string"
        description: "URL of the root resource where the content of the file specified\
          \ by the import job must be applied"
      status:
        type: "string"
        description: "Status of the import job (not started, running, succeeded, failed)"
      url:
        type: "string"
        description: "URL of the file containing the data to be imported"
      completionDate:
        type: "string"
        format: "date-time"
        description: "Date at which the job was completed"
      creationDate:
        type: "string"
        format: "date-time"
        description: "Date at which the job was created"
      errorLog:
        type: "string"
        description: "Reason for failure if status is failed"
    description: "Represents a task used to import resources from a file"
    example:
      path: "path"
      completionDate: "2000-01-23T04:56:07.000+00:00"
      errorLog: "errorLog"
      id: "id"
      href: "href"
      creationDate: "2000-01-23T04:56:07.000+00:00"
      contentType: "contentType"
      url: "url"
      status: "status"
  ImportJob_Create:
    type: "object"
    required:
    - "url"
    properties:
      contentType:
        type: "string"
        description: "Indicates the format of the imported data"
      path:
        type: "string"
        description: "URL of the root resource where the content of the file specified\
          \ by the import job must be applied"
      status:
        type: "string"
        description: "Status of the import job (not started, running, succeeded, failed)"
      url:
        type: "string"
        description: "URL of the file containing the data to be imported"
      completionDate:
        type: "string"
        format: "date-time"
        description: "Date at which the job was completed"
      creationDate:
        type: "string"
        format: "date-time"
        description: "Date at which the job was created"
      errorLog:
        type: "string"
        description: "Reason for failure if status is failed"
    description: "Represents a task used to import resources from a file\nSkipped\
      \ properties: id,href"
    example:
      path: "path"
      completionDate: "2000-01-23T04:56:07.000+00:00"
      errorLog: "errorLog"
      creationDate: "2000-01-23T04:56:07.000+00:00"
      contentType: "contentType"
      url: "url"
      status: "status"
  ExportJob:
    type: "object"
    properties:
      id:
        type: "string"
        description: "Identifier of the export job"
      href:
        type: "string"
        description: "Reference of the export job"
      query:
        type: "string"
        description: "Used to scope the exported data"
      path:
        type: "string"
        description: "URL of the root resource acting as the source for streaming\
          \ content to the file specified by the export job"
      contentType:
        type: "string"
        description: "The format of the exported data"
      status:
        type: "string"
        description: "Status of the export job (not started, running, succeeded, failed)"
      url:
        type: "string"
        description: "URL of the file containing the data to be exported"
      completionDate:
        type: "string"
        format: "date-time"
        description: "Data at which the job was completed"
      creationDate:
        type: "string"
        format: "date-time"
        description: "Date at which the job was created"
      errorLog:
        type: "string"
        description: "Reason for failure"
    description: "Represents a task used to export resources to a file"
    example:
      path: "path"
      query: "query"
      completionDate: "2000-01-23T04:56:07.000+00:00"
      errorLog: "errorLog"
      id: "id"
      href: "href"
      creationDate: "2000-01-23T04:56:07.000+00:00"
      contentType: "contentType"
      url: "url"
      status: "status"
...
  Error:
    required:
    - "code"
    - "message"
    properties:
      code:
        type: "integer"
        description: "An integer coding the error type. This is given to caller so\
          \ he can translate them if required."
      message:
        type: "string"
        description: "A short localized string that describes the error."
      description:
        type: "string"
        description: "(optional) A long localized error description if needed. It\
          \ can contain precise information about which parameter is missing, or what\
          \ are the identifier acceptable values."
      infoURL:
        type: "string"
        description: "(optional) A URL to online documentation that provides more\
          \ information about the error."